Allow modifying layout in the *Stop callbacks
If the user wants to e.g. constrain positioning of items to a bounding
box, one way to do it is to force the positioning of items when the
drag or resize stops. There needs to be a way for these callbacks to
provide a new layout if needed, or the asynchronous setState() call will
override anything set from the outside.

The root of the problem is double bookkeeping where the state and
props both hold the layout, but that is a larger refactoring.